+++ src/sys/os/term/ansi/namespace/define.rs
@@ -14,9 +14,9 @@ use crate::__ansi_consts;
 /// Those without `_B` contain a string slice.
 ///
 /// Functions ending with `_B` return either an array or a byte slice.
-/// Functions without `_B` return either a [`StringNonul`] or [`&str`].
+/// Functions without `_B` return either a [`StringNonNul`] or [`&str`].
 ///
-/// [`StringNonul`]: crate::StringNonul
+/// [`StringNonNul`]: crate::StringNonNul
 ///
 /// # Constants and Methods
 /// ## Escape codes